Version 10.47 [March 29th 2017]
Bug Fixes:
Fixed error when trying to import invalid profile table.
Fixed issue which prevented actionbutton icons from updating.
Fixed error caused by attempt to skin non-existent options for Mac.
Fixed the "Darken Inactive" setting for stance bar.

Noteworthy Info:
The friendly nameplates in dungeons and raids are currently broken.
While in a dungeon or raid instance the default WoW nameplates are
supposed to be active. Unfortunately this is not the case when the ElvUI
NamePlates module is enabled. There is no easy fix for it, as the issue
is with how the nameplate module was initially written. For the time
being you will have to either play without friendly nameplates in those
situations, or disable the ElvUI NamePlates module and use a dedicated
addon for NamePlates. We apologize for the inconvenience.

Version 10.46 [March 28th 2017] (for patch 7.1.5 and 7.2)
Misc. Changes:
Implemented changes to support the new patch 7.2.
Made various skin tweaks according to latest FrameXML changes.
Enhanced the code which calculates Artifact Power in bags for the
Artifact DataBar.

Noteworthy Info:
Because of a change Blizzard made to friendly nameplates in 7.2, it is
no longer possible to modify them in dungeons and raid instances. If you
use friendly nameplates in those situations then you will notice that
they use the default WoW style. There is nothing we can do about that.
More info here: https://eu.battle.net/forums/en/wow/topic/17615133023